[UPUP-1077]: fix comments
