+++ Example/Sources/Repository/PostsRepository.swift
@@ -1,5 +1,5 @@
 //
-//  PostRepository.swift
+//  PostsRepository.swift
 //  Example
 //
 //  Created by Maria Nesterova on 16.04.2025.
@@ -7,32 +7,18 @@
 
 import Foundation
 
-enum PostRepositoryError: Swift.Error {
-    case undefined
-}
-
-extension PostRepositoryError: LocalizedError {
-    var errorDescription: String? {
-        switch self {
-        case .undefined:
-            return "Ooops:("
-        }
-    }
-}
-
-final class PostRepository: Sendable {
+final class PostsRepository: Sendable {
     private enum Constants {
         static let delayInNanoseconds: UInt64 = 3_000_000_000
     }
     
-    @Sendable
     func getPosts(page: Int, pageSize: Int) async throws -> PostExampleModel {
         try await Task.sleep(nanoseconds: Constants.delayInNanoseconds)
         
         if let postExampleModel = getPostExampleData(page: page, pageSize: pageSize) {
             return postExampleModel
         } else {
-            throw PostRepositoryError.undefined
+            throw PostsRepositoryError.undefined
         }
     }
