+++ Example/Sources/UI/ContentView.swift
@@ -25,7 +25,7 @@ struct ContentView: View {
                 Button {
                     navigationPath.append(.listWithSection)
                 } label: {
-                    Text("Tap to go list with section")
+                    Text("Tap to go to list with section")
                         .padding(20)
                 }
                 .background(.gray)
@@ -34,7 +34,7 @@ struct ContentView: View {
                 Button {
                     navigationPath.append(.listWithoutSection)
                 } label: {
-                    Text("Tap to go list without section")
+                    Text("Tap to go to list without section")
                         .padding(20)
                 }
                 .background(.gray)
@@ -43,7 +43,7 @@ struct ContentView: View {
                 Button {
                     navigationPath.append(.listWithPageRequestService)
                 } label: {
-                    Text("Tap to go list with PageRequestService")
+                    Text("Tap to go to list with PageRequestService")
                         .padding(20)
                 }
                 .background(.gray)
