+++ Sources/PagingList/Extension/NotificationName+Ext.swift
@@ -1,0 +1,12 @@
+//
+//  NotificationName+Ext.swift
+//  PagingList
+//
+//  Created by Maria Nesterova on 22.04.2025.
+//
+
+import Foundation
+
+extension Notification.Name {
+    static let stopPrefetching = Notification.Name("StopPrefetching")
+}
